105 Construct Binary Tree from Preorder and Inorder
We use the preorder traversal to identify the current root of our
subtree. We then use the inorder traversal to determine the size of the
left subtree. Using this number, we can then create the correct call to
construct the left and right subtree.
For the left subtree, we skip the current root and use the part of the
preorder traversal that has the length of the left subtree. From the
inorder we extract until before the point, where the current root would
be.
For the right subtree, we skip the whole left subtree + current root in
the preorder and inorder traversals respectively.

Because we are slicing slices and slices are implemented via pointers,
we are not copying any memory with our go based implementation and thus
reach maximal efficiency of the O(|preorder|).