124 Binary Tree Max Path Sum
Because we are looking for a path, we must be able to walk this path
without visiting any node twice, while visiting all nodes.
This means if we start some path at some root, then whenever we extend
the path, we must choose the direction left or right, but cannot use
both.

The max path sum of a binary tree defined as root can be described as,
max {
    max path starting at the root,
    max path starting at the left subtree, if it exists
    max path starting at the right subtree, if it exists
}

the max path starting at some root is defined as,
root.Val + continuePath(root.Left) + continuePath(root.Right)
where continuePath(node) is defined as the max of
1. end -> return 0
2. continue left -> return node.Val + continuePath(node.Left)
3. continue right -> return node.Val + continuePath(node.Right)

We observe that continuePath(node) will not change for any node in the
tree, regardless of where exactly its path originated from.
Thus, we will call continuePath(node) for every node that is not a
direct descendent of the tree's root at least twice, expecting the exact
same result. This means we can cache the result for each node, and use
it in subsequent traversals, when we check a different starting root of
a path.