208 Implement trie
It is easiest to implement a trie recursively as it is a recursive data
structure.
It is important to mark that an inserted word ends at some node, in
order to properly decide the result of Search(...).
The adjacency list should be
Children map[rune]*Trie
Thus, it is not required to store the value of each node in Trie, as it
is already defined in its parent adjacency list.
The first rune and its size of a string can be extracted using,
start, size := utf8.DecodeRuneInString(word)